Move host isolation from eval-time to provisioning
Host firewall isolation was applied at sample_init by FirewallCommands talking
to the Proxmox REST API. That path could only infer the management interface
from the network config, which is ambiguous for DHCP/bridged-management hosts
and silently mis-fired (verified live: on a metal/slirp host it picked the
bridge port and locked the caller out). Isolation is also infrastructure state
configured once when the host is built, not a per-sample concern.

Provisioning now owns it. The build scripts in this repo run on the node with
root, pvesh, and the live routing table — the ground truth for which interface
external API/SSH traffic arrives on — so detection becomes a simple
`ip route` lookup with no heuristic.

- Remove the eval-time subsystem: delete FirewallCommands, the HostIsolation
  config model, the sample_init call, the InfraCommands wiring, and the two
  firewall test modules.
- Add scripts/configure_host_isolation.sh: a self-contained, idempotent
  reference that enables the cluster + node firewall, accepts the management
  ports (8006, 22) only on the default-route interface, and keeps SDN DNS/DHCP
  open. Run it on any node provisioned by other means.
- Wire the same logic into both provisioning paths: the on-first-boot heredoc
  in build_proxmox_auto.sh (runs once under slirp -> vmbr0, persists into every
  vend clone) and a new per-boot proxmox-ami-fixup-firewall.service in
  ec2/userdata.sh (detects the NIC per launch, like the existing fixup-nat
  service, so it survives AMI relaunch on differing instance families).
- Document the responsibility in README and CONTRIBUTING.
- Replace the deleted e2e test with a skippable one
  (INSPECT_PROXMOX_EXPECT_ISOLATION=1) that asserts a sandbox VM can't reach
  pveproxy/SSH on a provisioned-isolated host.

Verified end-to-end on a live host: the reference script detects enp39s0,
applies the rules idempotently, and the API caller stays reachable; with the
host isolated, a sandbox VM brought up via sample_init (now with no firewall
code in the eval path) cannot reach pveproxy or SSH while DNS/egress work.
The ec2/userdata.sh NitroVirt path needs a fresh AMI build to take effect.
